+++ src/test/java/com/coniverse/dangjang/domain/auth/service/OauthLoginServiceTest.java
@@ -16,6 +16,7 @@ import com.coniverse.dangjang.domain.auth.dto.AuthToken.AuthTokens;
 import com.coniverse.dangjang.domain.auth.dto.OauthProvider;
 import com.coniverse.dangjang.domain.auth.dto.Response.LoginResponse;
 import com.coniverse.dangjang.domain.auth.dto.request.KakaoLoginParams;
+import com.coniverse.dangjang.domain.auth.service.authToken.AuthTokensGenerator;
 import com.coniverse.dangjang.domain.user.dto.UserInfo;
 import com.coniverse.dangjang.domain.user.entity.User;
 import com.coniverse.dangjang.domain.user.infrastructure.KakaoInfoResponse;
