+++ src/test/java/com/coniverse/dangjang/global/config/TestConfig.java
@@ -5,10 +5,10 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
-import com.coniverse.dangjang.domain.auth.service.AuthTokensGenerator;
-import com.coniverse.dangjang.domain.auth.service.JwtTokenProvider;
 import com.coniverse.dangjang.domain.auth.service.OauthLoginService;
 import com.coniverse.dangjang.domain.auth.service.TestOauthInfoService;
+import com.coniverse.dangjang.domain.auth.service.authToken.AuthTokensGenerator;
+import com.coniverse.dangjang.domain.auth.service.authToken.JwtTokenProvider;
 import com.coniverse.dangjang.domain.user.repository.UserRepository;
 import com.coniverse.dangjang.domain.user.service.UserService;
